pyproject.toml
[project]
name = "adk-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "fastapi>=0.115.13",
    "google-adk>=1.3.0",
    "httpx>=0.28.1",
    "litellm>=1.72.6",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
]
